change: refactor GroupFeedController
